refactor(control): split connection handoff
