+++ crates/honk-core/src/control/connection/flow.rs
@@ -1,43 +1,5 @@
 use super::*;
 
-/// Build the eBPF conntrack key for a flow: IPs as 16-byte v4-mapped
-/// addresses, ports in host byte order, `l4proto` as the IANA number.
-pub(crate) fn build_tuples_key(
-    dst_ip: std::net::IpAddr,
-    dst_port: u16,
-    src_ip: std::net::IpAddr,
-    src_port: u16,
-    l4proto: u8,
-) -> TuplesKey {
-    // mem::zeroed, NOT TuplesKey::default(): the struct has 3 implicit
-    // padding bytes after l4proto (37 field bytes in a 40-byte repr(C)
-    // layout), and Rust does not guarantee padding is zeroed on field-wise
-    // initialization.  The kernel hashes all 40 key bytes, and the datapath
-    // writes keys from a zeroed scratch buffer — a garbage-padded userspace
-    // key never matches (lookups/deletes silently ENOENT).
-    let mut key: TuplesKey = unsafe { std::mem::zeroed() };
-    match dst_ip {
-        std::net::IpAddr::V4(ip) => {
-            key.dst_ip[10] = 0xff;
-            key.dst_ip[11] = 0xff;
-            key.dst_ip[12..16].copy_from_slice(&ip.octets());
-        }
-        std::net::IpAddr::V6(ip) => key.dst_ip.copy_from_slice(&ip.octets()),
-    }
-    match src_ip {
-        std::net::IpAddr::V4(ip) => {
-            key.src_ip[10] = 0xff;
-            key.src_ip[11] = 0xff;
-            key.src_ip[12..16].copy_from_slice(&ip.octets());
-        }
-        std::net::IpAddr::V6(ip) => key.src_ip.copy_from_slice(&ip.octets()),
-    }
-    key.dst_port = dst_port;
-    key.src_port = src_port;
-    key.l4proto = l4proto;
-    key
-}
-
 #[cfg(any(feature = "ebpf", test))]
 fn final_udp_rule_mark(routed_direct: bool, final_outbound: &str, routed_mark: u32) -> u32 {
     if final_outbound == "direct" && !routed_direct {
@@ -48,173 +10,6 @@ fn final_udp_rule_mark(routed_direct: bool, final_outbound: &str, routed_mark: u
 }
 
 impl ControlPlaneHandle {
-    /// `/proc/<pid>/exe` is display-only enrichment. Never hold first-packet
-    /// delivery behind the blocking pool used to resolve it.
-    fn spawn_process_path_enrichment(&self, conn_id: String, handoff: Option<&HandoffResult>) {
-        let Some(handoff) = handoff.filter(|handoff| handoff.pid != 0).cloned() else {
-            return;
-        };
-        let tracker = Arc::clone(&self.connection_tracker);
-        tokio::spawn(async move {
-            if let Some(process_path) = handoff.process_path().await {
-                tracker.update_process_path(&conn_id, process_path);
-            }
-        });
-    }
-    /// Look up the eBPF routing handoff entry for a connection, consuming it.
-    ///
-    /// Only a read lock is taken: `routing_handoff_take` performs raw bpf()
-    /// map operations, which the kernel serializes internally — no userspace
-    /// backend state is touched.  The lock's sole role here is to keep the
-    /// backend (and its map fds) alive against `cleanup()`, which takes the
-    /// write lock.
-    async fn lookup_handoff(&self, tuples: &TuplesKey) -> Option<HandoffResult> {
-        self.ebpf
-            .read()
-            .await
-            .routing_handoff_take(tuples)
-            .ok()
-            .flatten()
-            .map(Into::into)
-    }
-
-    /// Staged UDP transitions consume their handoff atomically at commit, so
-    /// initialization may only inspect it. Legacy socket ingress keeps the
-    /// existing take-once behavior.
-    async fn lookup_udp_handoff(
-        &self,
-        tuples: &TuplesKey,
-        decision_token: u32,
-    ) -> anyhow::Result<Option<HandoffResult>> {
-        if decision_token == 0 {
-            return Ok(self.lookup_handoff(tuples).await);
-        }
-        let entry = self
-            .ebpf
-            .read()
-            .await
-            .routing_handoff_lookup(tuples)?
-            .ok_or_else(|| anyhow::anyhow!("staged UDP flow has no routing handoff"))?;
-        if entry.result.decision_token != decision_token {
-            anyhow::bail!(
-                "staged UDP handoff token mismatch: expected {}, found {}",
-                decision_token,
-                entry.result.decision_token
-            );
-        }
-        Ok(Some(entry.into()))
-    }
-
-    async fn adopt_tcp_flow(
-        &self,
-        stream: TcpStream,
-        tuples: TuplesKey,
-    ) -> anyhow::Result<(TcpFlowGuard, Option<HandoffResult>)> {
-        let backend = self.ebpf.read().await;
-        match backend.tcp_conn_state_lookup(&tuples) {
-            Ok(Some(_)) => {}
-            Ok(None) => anyhow::bail!("accepted TCP flow has no conn-state: {tuples:?}"),
-            Err(error) => {
-                return Err(anyhow::anyhow!(
-                    "accepted TCP flow conn-state lookup failed for {tuples:?}: {error}"
-                ));
-            }
-        }
-
-        let flow = TcpFlowGuard::new(
-            stream,
-            tuples,
-            Arc::clone(&self.tcp_flow_pins),
-            Arc::clone(&self.ebpf),
-            Arc::clone(&self.connection_tracker),
-        );
-        let handoff = backend
-            .routing_handoff_take(&tuples)
-            .ok()
-            .flatten()
-            .map(Into::into);
-        Ok((flow, handoff))
-    }
-
-    async fn outbound_index_to_name(&self, index: u8) -> String {
-        match OutboundIndex::from_user(index as u32) {
-            OutboundIndex::Direct => "direct".into(),
-            OutboundIndex::Block => "block".into(),
-            OutboundIndex::MustRules => "must_rules".into(),
-            OutboundIndex::ControlPlaneRouting => "control_plane_routing".into(),
-            _ => {
-                let config = self.config.read().await;
-                // Map user index back to the group name (same order as
-                // outbound_name_to_id above).
-                let user_idx = index.saturating_sub(OutboundIndex::UserBase as u8);
-                config
-                    .groups
-                    .get(user_idx as usize)
-                    .map(|g| g.name.clone())
-                    .unwrap_or_else(|| config.routing.default_outbound.clone())
-            }
-        }
-    }
-
-    #[cfg(feature = "ebpf")]
-    async fn outbound_name_to_index(&self, outbound_name: &str) -> u8 {
-        match outbound_name {
-            "direct" => OutboundIndex::Direct as u8,
-            "block" => OutboundIndex::Block as u8,
-            "must_rules" => OutboundIndex::MustRules as u8,
-            "control_plane_routing" => OutboundIndex::ControlPlaneRouting as u8,
-            _ => {
-                let config = self.config.read().await;
-                config
-                    .groups
-                    .iter()
-                    .position(|group| group.name == outbound_name)
-                    .and_then(|index| u8::try_from(index).ok())
-                    .and_then(|index| (OutboundIndex::UserBase as u8).checked_add(index))
-                    .unwrap_or(OutboundIndex::ControlPlaneRouting as u8)
-            }
-        }
-    }
-
-    /// Clash mode override (approximate clash semantics), applied after the
-    /// eBPF handoff / userspace Router produced an outbound and before
-    /// `resolve_outbound_nodes`:
-    ///
-    /// - mode `Direct` forces `direct`;
-    /// - mode `Global` forces the current GLOBAL selection (a group or node
-    ///   name, resolved via the normal path; when it resolves to nothing the
-    ///   original routing result is kept);
-    /// - `block` results and `must` results (dae `(must)` rules / eBPF
-    ///   handoff must flag) are never overridden — both are final routing
-    ///   decisions that mode switches must not bypass.
-    async fn apply_mode_override(&self, outbound_name: String, must: bool) -> String {
-        let Some(ref mode_state) = self.mode_state else {
-            return outbound_name;
-        };
-        if must || outbound_name == "block" {
-            return outbound_name;
-        }
-        let state = { mode_state.read().clone() };
-        // The GLOBAL selection needs a config lookup to decide whether it
-        // resolves to a group/node; only do it in Global mode.
-        let mut selection_resolvable = false;
-        if state.is_global() && !state.global_selection.is_empty() {
-            let selection = &state.global_selection;
-            selection_resolvable = *selection == "direct" || *selection == "block" || {
-                let config = self.config.read().await;
-                config.groups.iter().any(|g| g.name == *selection)
-                    || config.nodes.iter().any(|n| n.name == *selection)
-            };
-            if !selection_resolvable {
-                debug!(
-                    "clash Global selection '{}' does not resolve; keeping routed outbound '{}'",
-                    selection, outbound_name
-                );
-            }
-        }
-        state.override_outbound(&outbound_name, false, selection_resolvable)
-    }
-
     pub(in crate::control) async fn serve_connection(
         &self,
         stream: TcpStream,
@@ -1964,6 +1759,3 @@ mod dial_permit_scope_tests;
 #[cfg(test)]
 #[path = "sniffed_domain_routing_tests.rs"]
 mod sniffed_domain_routing_tests;
-#[cfg(test)]
-#[path = "tcp_flow_lifecycle_tests.rs"]
-mod tcp_flow_lifecycle_tests;
